Renamed base package name for StandardizedAnalysisAPI
