+++ src/main/java/org/ohdsi/analysis/Cohort.java
@@ -1,4 +1,4 @@
-package org.ohdsi.standardized_analysis_api;
+package org.ohdsi.analysis;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import org.ohdsi.circe.cohortdefinition.CohortExpression;
