+++ src/main/java/org/ohdsi/analysis/cohortcharacterization/design/FeatureAnalysisType.java
@@ -1,4 +1,4 @@
-package org.ohdsi.standardized_analysis_api.cohortcharacterization.design;
+package org.ohdsi.analysis.cohortcharacterization.design;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
