+++ src/main/java/org/ohdsi/analysis/cohortcharacterization/result/BaseStat.java
@@ -1,4 +1,4 @@
-package org.ohdsi.standardized_analysis_api.cohortcharacterization.result;
+package org.ohdsi.analysis.cohortcharacterization.result;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
